+++ callbackquery_test.go
@@ -22,6 +22,9 @@ func TestRefreshCallbackHandler(t *testing.T) {
 		return now
 	}
 	dm, err := NewMockBusArrivalAPI(now)
+	if err != nil {
+		t.Fatal(err)
+	}
 	defer dm.Close()
 	datamallEndpoint = dm.URL