+++ command_test.go
@@ -127,6 +127,9 @@ func TestEtaHandler(t *testing.T) {
 		return now
 	}
 	dm, err := NewMockBusArrivalAPI(now)
+	if err != nil {
+		t.Fatal(err)
+	}
 	defer dm.Close()
 	datamallEndpoint = dm.URL