+++ inlinequery_test.go
@@ -132,6 +132,9 @@ func TestChosenInlineResultHandler(t *testing.T) {
 		return now
 	}
 	dm, err := NewMockBusArrivalAPI(now)
+	if err != nil {
+		t.Fatal(err)
+	}
 	defer dm.Close()
 	datamallEndpoint = dm.URL